requirements.txt
Flask==2.1.2
Flask-RESTful==0.3.9
python-dateutil==2.8.2
python-twitch-client==0.7.1
requests==2.27.1
gunicorn==20.1.0